--- core/src/main/java/de/fraunhofer/iosb/ilt/faaast/client/util/HttpFactory.java
@@ -1,57 +1,0 @@
-/*
- * Copyright (c) 2024 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
- * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
- * Forschung e.V.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package de.fraunhofer.iosb.ilt.faaast.client.util;
-
-import java.net.http.HttpClient;
-import javax.net.ssl.*;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.X509Certificate;
-
-
-/**
- * Custom Factory for trusting self signed certs.
- */
-public class HttpFactory {
-    /**
-     * Creates the http client with trust.
-     *
-     * @return the http client
-     * @throws NoSuchAlgorithmException when SSL algorithm is wrong
-     * @throws KeyManagementException when key issues arise
-     */
-    public static HttpClient createHttpClient() throws NoSuchAlgorithmException, KeyManagementException {
-        SSLContext sslContext = SSLContext.getInstance("TLS");
-        sslContext.init(null, new TrustManager[] {
-                new X509TrustManager() {
-                    @Override
-                    public void checkClientTrusted(X509Certificate[] certs, String authType) {}
-
-
-                    @Override
-                    public void checkServerTrusted(X509Certificate[] certs, String authType) {}
-
-
-                    @Override
-                    public X509Certificate[] getAcceptedIssuers() {
-                        return new X509Certificate[0];
-                    }
-                }
-        }, new java.security.SecureRandom());
-        return HttpClient.newBuilder()
-                .sslContext(sslContext)
-                .build();
-    }
-}
